# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple delivery app API",
        "title": "Delivery App API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/api/productCategories": {
            "get": {
                "description": "Returns all product categories from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "productCategories"
                ],
                "summary": "Get list of product categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProductCategory"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/productCategory": {
            "post": {
                "description": "Adds a new product category to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "productCategories"
                ],
                "summary": "Create a new product category",
                "parameters": [
                    {
                        "description": "ProductCategory data",
                        "name": "productCategory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ProductCategory created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/productCategory/{id}": {
            "get": {
                "description": "Returns a product category by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "productCategories"
                ],
                "summary": "Get product category by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Updates product category information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "productCategories"
                ],
                "summary": "Update a product category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated product category data",
                        "name": "productCategory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product category updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Product category not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a product category by ID",
                "tags": [
                    "productCategories"
                ],
                "summary": "Delete a product category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product category deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Product category not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ProductCategory": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "statusId": {
                    "type": "integer"
                }
            }
        },
...
    }
}